pyproject.toml
[project]
name = "mindrian-agno"
version = "1.0.0"
description = "Mindrian AI-powered innovation platform built on Agno framework"
requires-python = ">=3.11"
dependencies = [
    # ─────────────────────────────────────────────────────────────────────────
    # Core Framework
    # ─────────────────────────────────────────────────────────────────────────
    "agno>=1.0.0",

    # ─────────────────────────────────────────────────────────────────────────
    # LLM Providers (Multi-provider: Anthropic + Google)
    # ─────────────────────────────────────────────────────────────────────────
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",

    # ─────────────────────────────────────────────────────────────────────────
    # Database Clients
    # ─────────────────────────────────────────────────────────────────────────
    "neo4j>=5.0.0",                  # Knowledge graph
    "supabase>=2.0.0",               # Conversation memory + pgvector
    "sqlalchemy>=2.0.0",             # ORM for local storage

    # ─────────────────────────────────────────────────────────────────────────
    # MCP (Model Context Protocol)
    # ─────────────────────────────────────────────────────────────────────────
    "mcp>=1.0.0",                    # MCP client library

    # ─────────────────────────────────────────────────────────────────────────
    # API Framework
    # ─────────────────────────────────────────────────────────────────────────
    "fastapi[standard]>=0.115.0",
    "uvicorn>=0.32.0",
    "httpx>=0.27.0",                 # Async HTTP client
    "websockets>=12.0",              # WebSocket support

    # ─────────────────────────────────────────────────────────────────────────
    # Configuration & Validation
    # ─────────────────────────────────────────────────────────────────────────
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.0",

    # ─────────────────────────────────────────────────────────────────────────
    # Utilities
    # ─────────────────────────────────────────────────────────────────────────
    "rich>=13.0.0",                  # Terminal formatting
    "structlog>=24.0.0",             # Structured logging
    "tenacity>=9.0.0",               # Retry logic
    "jinja2>=3.1.0",                 # Template engine for code generation
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

ui = [
    "streamlit>=1.40.0",
    "chainlit>=1.3.0",
]

all = [
    "mindrian-agno[dev,ui]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"
